lg_app/rm_agent_app/pyproject.toml
[project]
name = "rm-agent"
version = "0.1.0"
description = "Recipe Manager AI Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grandalf>=0.8",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
